chore: move nextjs config to the root as it's copy-paste in every project
