+++ apps/events-helsinki/src/pages/500.tsx
@@ -1,27 +1,15 @@
-import {
-  getLanguageOrDefault,
-  UnknownError,
-  useCommonTranslation,
-} from 'events-helsinki-components';
-
-import type { GetStaticPropsContext, NextPage } from 'next';
+import { useRouter } from 'next/router';
 import React from 'react';
-import getEventsStaticProps from '../domain/app/getEventsStaticProps';
-import serverSideTranslationsWithCommon from '../domain/i18n/serverSideTranslationsWithCommon';
 
-const FiveHundred: NextPage = () => {
-  const { t } = useCommonTranslation();
-  return <UnknownError appName={t(`appSports:appName`)} />;
+/**
+ * Due to middleware bug: https://github.com/vercel/next.js/issues/38762
+ * let's redirect always to the pages/error.tsx page.
+ */
+const NextErrorPage = () => {
+  const router = useRouter();
+  React.useEffect(() => {
+    router.push('/error');
+  }, [router]);
+  return null;
 };
-export default FiveHundred;
-
-export async function getStaticProps(context: GetStaticPropsContext) {
-  return getEventsStaticProps(context, async () => {
-    const language = getLanguageOrDefault(context.locale);
-    return {
-      props: {
-        ...(await serverSideTranslationsWithCommon(language, ['common'])),
-      },
-    };
-  });
-}
+export default NextErrorPage;
