+++ apps/hobbies-helsinki/package.json
@@ -92,7 +92,6 @@
     "@graphql-codegen/typescript-operations": "2.4.1",
     "@graphql-codegen/typescript-react-apollo": "3.2.15",
     "@graphql-codegen/typescript-resolvers": "^2.6.5",
-    "@next/bundle-analyzer": "13.1.6",
     "@next/env": "13.1.6",
     "@size-limit/file": "8.1.0",
     "@storybook/react": "6.5.12",
