+++ package.json
@@ -74,7 +74,12 @@
     "postinstall": "is-ci || yarn husky install"
   },
   "dependencies": {
-    "react": "18.2.0"
+    "@next/bundle-analyzer": "12.3.1",
+    "@sentry/nextjs": "7.14.1",
+    "next": "12.3.1",
+    "next-transpile-modules": "9.0.0",
+    "react": "18.2.0",
+    "react-dom": "18.2.0"
   },
   "devDependencies": {
     "@babel/core": "7.20.7",
@@ -114,6 +119,7 @@
     "nodemon": "2.0.20",
     "npm-run-all": "4.1.5",
     "path": "0.12.7",
+    "picocolors": "1.0.0",
     "prettier": "2.7.1",
     "rimraf": "3.0.2",
     "shell-quote": "1.7.3",
